requirements_gui.txt
gradio>=3.0.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
numpy>=1.21.0
PyYAML>=5.4.0
natsort>=8.0.0
kornia>=0.6.0
einops>=0.4.0
timm>=0.6.0
torchmetrics>=0.9.0

